+++ BackgroundSubtractorPBAS_1ch.h
@@ -12,10 +12,10 @@
 class BackgroundSubtractorPBAS_1ch : public BackgroundSubtractorPBAS {
 public:
 	//! full constructor
-	BackgroundSubtractorPBAS_1ch(	int nInitColorDistThreshold=BGSPBAS_DEFAULT_COLOR_DIST_THRESHOLD,
+	BackgroundSubtractorPBAS_1ch(	size_t nInitColorDistThreshold=BGSPBAS_DEFAULT_COLOR_DIST_THRESHOLD,
 									float fInitUpdateRate=BGSPBAS_DEFAULT_LEARNING_RATE,
-									int nBGSamples=BGSPBAS_DEFAULT_NB_BG_SAMPLES,
-									int nRequiredBGSamples=BGSPBAS_DEFAULT_REQUIRED_NB_BG_SAMPLES);
+									size_t nBGSamples=BGSPBAS_DEFAULT_NB_BG_SAMPLES,
+									size_t nRequiredBGSamples=BGSPBAS_DEFAULT_REQUIRED_NB_BG_SAMPLES);
 	//! default destructor
 	virtual ~BackgroundSubtractorPBAS_1ch();
 	//! (re)initiaization method; needs to be called before starting background subtraction
